legacy file upload render simple details
This commit enables the simple hash project index during the project
upload file. This change will store the simple storage with the
project hash index for new project uploads.

This PR is a follow-up on PR #8586, where it uses in the legacy file
upload the render simple details as part of PEP 458.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>