Manage packages/project and simple details to TUF
* Adding packages

After adding a package to the Warehouse database, it generates and
stores the Simple Index with a request to the RSTUF backend to
include the package and its simple index in TUF Metadata.

* Removing package or Project Release

On PyPI Management, when a user removes a file or a project release
it also removes it from TUF metadata and updates the simple details index.

Co-authored-by: Lukas Puehringer <lukas.puehringer@nyu.edu>
Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>

simplify code in warehouse.tuf.targets

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>